+++ modules/impex/core/src/sap/commerce/toolset/impex/codeInsight/completion/provider/ImpExMacrosCompletionProvider.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -41,6 +41,7 @@ class ImpExMacrosCompletionProvider : CompletionProvider<CompletionParameters>()
             .map { ImpExMacroReference.escapeName(it.text) }
             .map { ImpExLookupElementFactory.buildMacro(it) }
             .let { result.addAllElements(it) }
-    }
 
+        result.addElement(ImpExLookupElementFactory.buildMacroConfig())
+    }
 }