+++ modules/impex/core/src/sap/commerce/toolset/impex/lang/refactoring/ImpExRefactoringSupportProvider.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -25,5 +25,5 @@ import sap.commerce.toolset.impex.psi.ImpExPsiNamedElement
 class ImpExRefactoringSupportProvider : RefactoringSupportProvider() {
 
     override fun isMemberInplaceRenameAvailable(element: PsiElement, context: PsiElement?) = element is ImpExPsiNamedElement
-        && !element.text.startsWith(ImpExConstants.IMPEX_CONFIG_COMPLETE_PREFIX)
+        && !element.text.startsWith(ImpExConstants.MACRO_CONFIG_COMPLETE_MARKER)
 }